refactor(persistence): ADR-0035 slice 4 — extract verifier_store into the kirra-persistence crate
The payoff of the enabling slices (1–3.5, all merged): verifier_store named zero
root-crate items, so the wholesale move is mechanical. `git mv` the whole module
dir into crates/kirra-persistence (mod.rs -> lib.rs); rewrite the
`crate::verifier_store::` self-refs -> `crate::`; src/verifier_store.rs becomes a
`pub use kirra_persistence::*;` shim so every `crate::verifier_store::*` /
`kirra_verifier::verifier_store::*` path (service binary, kirra-verifier-pg backend,
tests) resolves unchanged. The crate depends ONLY on the lean leaf crates
(kirra-core / kirra-audit-hash / kirra-fleet-types / kirra-ota-campaign /
kirra-fabric-types) + storage/crypto externals — never back on the verifier tree.

Two couplings that live OUTSIDE verifier_store/ (so the enabling audit hadn't
reached them), resolved here:
- `impl FleetTrustStore for VerifierStore` (src/fleet_trust_store.rs): both types
  are now external to root -> orphan rule. Relocated into kirra-persistence (owns
  VerifierStore); its fleet-role key resolution is inlined (a base64 decode of the
  stored controller key) so it needs neither the root KeyRegistry nor
  crate::attestation. This orphaned KeyRegistry in root (its sole in-root consumer);
  baselined with a follow-up note (relocate-or-delete).
- `#[cfg(test)]` store helpers (`*_for_test`, the durable-write fault-injection
  fields + write-path hooks) that ROOT tests call: a dependency compiles without its
  own test cfg, so a `test-support` feature on kirra-persistence gates them; the root
  crate's dev-dependency enables it. Off for production builds.

Root's audit_chain delegates to `kirra_persistence::append_audit_event_tx` (now
`pub` — a legitimate cross-crate API, resolving the #928 review note).

Behaviour-preserving, byte-identical, test-count-preserving. Verified: full
workspace build; root lib 607 + kirra-persistence 153 = the prior 760; power-loss
audit drill (3), two_node_rollout (4), ha_failover (4); kirra-verifier-pg builds
through the shim; fmt, guardrails, orphan-cores green. Detached lockfiles (fuzz /
kirra-verifier-pg / parko) refreshed for the new crate.
