+++ Cargo.toml
@@ -19,7 +19,7 @@
 # runtime (ORT etc.) is absent, strict-fail only in the CI lane that installs
 # it (see parko-onnx's PARKO_ONNX_REQUIRE_ORT).
 [workspace]
-members = [".", "crates/kirra-core", "crates/kirra-policy-types", "crates/kirra-industrial", "crates/kirra-contract-channel", "crates/kirra-hv-carrier", "crates/kirra-inline-governor", "crates/kirra-l3-e2e", "crates/kirra-release-token", "crates/kirra-timing", "crates/kirra-wcet-bench", "crates/kirra-trajectory", "crates/kirra-fleet-types", "crates/kirra-ros2-adapter", "crates/kirra-capture-schema", "crates/kirra-collector", "crates/kirra-governor-service", "crates/kirra-wire-client", "crates/kirra-proposal-bench", "crates/kirra-planner", "crates/kirra-mick", "crates/kirra-map", "crates/kirra-taj", "crates/kirra-fleet-transport", "crates/kirra-doer-eval", "crates/kirra-kpi-gate", "crates/kirra-ota-installer", "crates/kirra-ota-campaign", "crates/kirra-fabric-types", "crates/kirra-audit-hash", "crates/kirra-replay", "crates/kirra-actuation-consumer", "crates/kirra-sidecars", "crates/kirra-consumer-ffi"]
+members = [".", "crates/kirra-core", "crates/kirra-policy-types", "crates/kirra-industrial", "crates/kirra-contract-channel", "crates/kirra-hv-carrier", "crates/kirra-inline-governor", "crates/kirra-l3-e2e", "crates/kirra-release-token", "crates/kirra-timing", "crates/kirra-wcet-bench", "crates/kirra-trajectory", "crates/kirra-fleet-types", "crates/kirra-ros2-adapter", "crates/kirra-capture-schema", "crates/kirra-collector", "crates/kirra-governor-service", "crates/kirra-wire-client", "crates/kirra-proposal-bench", "crates/kirra-planner", "crates/kirra-mick", "crates/kirra-map", "crates/kirra-taj", "crates/kirra-fleet-transport", "crates/kirra-doer-eval", "crates/kirra-kpi-gate", "crates/kirra-ota-installer", "crates/kirra-ota-campaign", "crates/kirra-fabric-types", "crates/kirra-audit-hash", "crates/kirra-persistence", "crates/kirra-replay", "crates/kirra-actuation-consumer", "crates/kirra-sidecars", "crates/kirra-consumer-ffi"]
 resolver = "2"
 
 [package]
@@ -114,6 +114,7 @@ kirra-release-token = { path = "crates/kirra-release-token", features = ["serde"
 kirra-ota-campaign = { path = "crates/kirra-ota-campaign" }
 kirra-fabric-types = { path = "crates/kirra-fabric-types" }
 kirra-audit-hash = { path = "crates/kirra-audit-hash" }
+kirra-persistence = { path = "crates/kirra-persistence" }
 parko-core = { path = "parko/crates/parko-core" }
 # Governor-free wire schema for capture records, shared verbatim with the offline
 # collector (docs/COLLECTOR_DESIGN.md [C1]). Re-exported by `crate::capture`.
@@ -161,6 +162,12 @@ hyper-util = { version = "0.1", features = ["server-auto", "tokio"] }
 [dev-dependencies]
 rand = "0.8"
 proptest = "1"
+# ADR-0035 slice 4: enable kirra-persistence's `test-support` feature for the root
+# crate's TEST build, so root tests (campaign_monitor, cert_expiry_monitor, the
+# posture/audit/HA suites) can call the VerifierStore `*_for_test` helpers that are
+# `#[cfg(test)]` inside the persistence crate (a dependency compiles without its own
+# test cfg). Off for the production `cargo build` (dev-deps don't apply there).
+kirra-persistence = { path = "crates/kirra-persistence", features = ["test-support"] }
 # EP-01: the WCET CI gate is extended to the IN-LINE loop (read → validate →
 # bound → sign → verify → release-decision) via the assembled crate's own API.
 kirra-inline-governor = { path = "crates/kirra-inline-governor" }
