+++ ci/orphan_cores_baseline.json
@@ -4,7 +4,7 @@
     "kirra_map::lanelet2": "Feature-gated Lanelet2 cxx FFI bridge (`lanelet2` feature, off by default); its consumer is the ros2+lanelet2 node build, outside the default-feature textual scan.",
     "kirra_planner::lanelet2": "Same feature-gated Lanelet2 bridge, planner side — consumed only under the `lanelet2` feature build.",
     "kirra_trajectory::redundancy_hardening": "Verification-hardening module by design: its consumers are the property/hardening test suites; it backs the True-Redundancy cross_check evidence, not a runtime path.",
-    "kirra_verifier::fleet_trust_store": "Trait-impl seam (impls kirra_fleet_types::store::FleetTrustStore for VerifierStore); consumption happens through the trait object at the fleet-transport integration boundary — invisible to a textual module-path scan.",
+    "kirra_verifier::key_registry": "Key-resolution facade (multi-role resolve + verify_ed25519_signature over a stored key). Its sole in-root consumer was the FleetTrustStore impl (src/fleet_trust_store.rs); at the ADR-0035 slice-4 persistence extraction that impl moved into kirra-persistence (orphan rule) and inlined its fleet-role resolution (a base64 decode), leaving KeyRegistry with no in-root consumer. TODO(follow-up): relocate KeyRegistry to kirra-persistence (needs attestation::parse_ed25519_public_pem relocated) or delete if it stays unused — kept this slice to hold the big mechanical move reviewable.",
     "kirra_verifier::health": "GENUINE ORPHAN (legacy DeploymentLivenessStatus helpers; the live /health route is implemented in the service bin). TODO: wire or delete — kept this cycle to hold the gate's introduction mechanical.",
     "kirra_verifier::traceability_gate": "CI gate by design: runs as a test lane (SAFETY:/REQ:/TEST: tag coverage); a runtime consumer would be wrong.",
     "kirra_verifier::wcet_gate": "CI gate by design: the O(1) structural-boundedness WCET guard runs as a test lane; a runtime consumer would be wrong.",
