+++ crates/kirra-persistence/src/audit_appender.rs
