+++ crates/kirra-persistence/src/fabric.rs
