+++ crates/kirra-persistence/src/federation.rs
