+++ crates/kirra-persistence/src/nodes.rs
