+++ crates/kirra-persistence/src/operators.rs
